Add some (simple) unit tests to make sure the results of the toolbox functions are maintained between changes.
At the moment, this is only a simple check against an existing result, so it is not particularly comprehensive, but it should sound alarm bells if something has broken catastrophically.

The data against which grid2fvcom is compared is not included in the toolbox as it is too large (and this whole testing thing might be moved to a separate repository if it makes more sense that way).