Update the NCEP script to actually use the NCEP Reanalysis-2 data (rather than the original reanalysis data).
The function now also downloads the geopotential height at the surface
used in the model. The original idea was that this would be able to mask
out the land data, but it turns out that the geopotential height does
not correlate with height above MSL very well, so it's difficult to mask
out the land areas. I've left the code in in case it's needed in the
future, for whatever reason.