Change the way the level is identified and the way the position grids are stored.
Make the search for the appropriate level slightly more precise (change the target pressure to 1013 from 1000 mbar, not that this will make any difference as the closest levels are 1000 and 9000).

Also added a more robust method for finding an appropriately sized array by finding the index which is not from relative humidity, topogrphy or pressure.