Change the function to allow download of either reanalysis-1, reanalysis-2 or 20th Century reanalysis 2 data.
Rather than have separate functions for each of the NCEP products, have
a single function which just downloads the relevant data depending on an
input argument.

So far, this supports the reanalysis-1, reanalysis-2 and 20th Century
reanalysis-2 data. All are the same resolution, but the variables are on
different grids (I think).

I've tested this with a combination of the three different sources and
with specific variables to extract, and it seems to work OK. I haven't
yet checked that the returned data from the 20th Century reanalysis-2
are sensible, but I'm pretty certain the reanalysis-1 and reanalysis-2
data are OK.