Change the function to use a URL for each time step.
Since the 4D variables are extracted a single time step at a time, the
URL which is used to get the data can be selected for each time step.
This means that this function is now able to extract data from across
year boundaries and from multiple different URLs. This required some
fairly major changes to the way the function was organised and needed
a small child function to select the appropriate URL.

The code has also seen some clean up (in terms of only doing things once
and not bothering with preallocating arrays which don't need to be
preallocated).