Fix the interpolation of the HYCOM data onto the vertical grid of FVCOM.
There's a bit of a problem with the interpolation in that sometimes the HYCOM data isn't properly picked up extending to its full depth. Howver, for my Irish Sea domain I've checked the interpolated values and things seem OK. Normalising the HYCOM depths to the FVCOM depths seems to account for most of the problems with under or overshooting HYCOM data.

I've also added various plots (commented out) for debugging. However, the large set of figures at the end of the figure don't work (I'm not sure why).