Fix the handling of boundary nodes which fall on land.
Create a land mask which is tested for each value in which the interpolation fails. If we're at the surface, then use the nearest value, otherwise leave the NaN at this depth. Later on when we're interpolating vertically, we check if we've only got one value, and if so, we just repeat it at all water depths. Otherwise we do the usual interpolation onto the FVCOM vertical layers.