Make a minor change to the vertical interpolation to support HYCOM data.
Since the HYCOM data are a fixed water depths rather than sigma levels,
we need to create an index of the values which are below the water depth
at this node and then eliminate them from the linear interpolation.
Fortunately this is relatively easy to do (so long as we create an
appropriate depth array) in such a manner as won't affect the POLCOMS
vertical interpolation.