Fix the net heat flux variable name and flip the latitude array to ensure the data are properly referenced in FVCOM.
It appears that for version 3.1.6 (official) of FVCOM, the expected
variable name in the WRF format netCDF file is not Net_heat (as Dima's
file has) but is instead net_heat_flux (as the rest of FVCOM uses).

The latitude variable needs to be flipped in the same way as the data
arrays, or, none need to be flipped. The downside to the latter is that
although FVCOM will read them correctly, viewing them with something
like ncview will render them upsidedown. Not a massive pain, but enough
of a nuisance for me when I was debugging this that it was easier to
flip everything.