New function to interpolate the fixed level HYCOM data onto the FVCOM horizontal and vertical grid.
Interpolate HYCOM's odd fixed level data onto the FVCOM model grid to
use as a warm start for FVCOM. During the coding of this function, it
became apparent that the interp_POLCOMS2FVCOM.m function is pretty
poorly written from a variable number of inputs perspective. This
function can probably be used as a template to improve the POLCOMS
version, should it be necessary.