Fix the way time is handled when reading in the newest Met Office model outputs (2012 onwards).
Previously a time variable had to be specified in varlist. Now, each
data variable's time is returned as an array within the MetUM.(variable)
struct, giving MetUM.(variable).time and MetUM.(variable).data. This
means if each data variable uses a different time sampling, that can be
accounted for later (by interpolating to a common time series with
interp3, for example).

Currently the code extracts the first 6 hour's worth of data. The
assumption there is that the Met Office do 4 runs per day, so 6 hours of
data from each run gives you a day's worth.