Change the way multiple rivers associated with a single model node are handled.
Instead of summing the discharges for each contributing river, use the
larger of the set.

This means that for the Fowey, England, for example, the discharge is
set at the larger of the two modelled time series instead of using the
sum of the two river time series. I had previously got very fresh
surface water due to the excessively large flux from the Fowey making
its way to the L4 buoy time series.